backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
cryptography==41.0.7
python-dotenv==1.0.0
